Synchronize selection
