+++ WalletWasabi.Fluent/ViewModels/CoinControl/CoinSelectorViewModel.cs
@@ -1,6 +1,8 @@
 using System.Collections.Generic;
 using System.Collections.Immutable;
+using System.Collections.ObjectModel;
 using System.Linq;
+using System.Reactive.Disposables;
 using System.Reactive.Linq;
 using Avalonia.Controls;
 using Avalonia.Controls.Models.TreeDataGrid;
@@ -9,30 +11,39 @@ using DynamicData;
 using NBitcoin;
 using ReactiveUI;
 using WalletWasabi.Blockchain.TransactionOutputs;
-using WalletWasabi.Fluent.Extensions;
 using WalletWasabi.Fluent.Helpers;
-using WalletWasabi.Fluent.Models;
 using WalletWasabi.Fluent.TreeDataGrid;
 using WalletWasabi.Fluent.ViewModels.CoinControl.Core;
 using WalletWasabi.Fluent.ViewModels.Wallets;
 using WalletWasabi.Fluent.Views.CoinControl.Core.Cells;
 using WalletWasabi.Fluent.Views.CoinControl.Core.Headers;
+using WalletWasabi.Wallets;
 
 namespace WalletWasabi.Fluent.ViewModels.CoinControl;
 
 public partial class CoinSelectorViewModel : ViewModelBase, IDisposable
 {
+	private readonly CompositeDisposable _disposables = new();
+	private readonly ObservableCollection<CoinControlItemViewModelBase> _source;
+	private readonly Wallet _wallet;
 	[AutoNotify] private IReadOnlyCollection<SmartCoin> _selectedCoins = ImmutableList<SmartCoin>.Empty;
+	private IDisposable _selectedCoinsUpdater = Disposable.Empty;
 
-	public CoinSelectorViewModel(WalletViewModel walletViewModel, IEnumerable<SmartCoin> initialCoinSelection)
+	public CoinSelectorViewModel(WalletViewModel walletViewModel, IList<SmartCoin> initialCoinSelection)
 	{
-		var pockets = walletViewModel.Wallet.GetPockets();
-		var pocketItems = CreatePocketItems(pockets);
+		_wallet = walletViewModel.Wallet;
+		_source = new ObservableCollection<CoinControlItemViewModelBase>();
 
-		SyncSelectedItems(pocketItems, initialCoinSelection);
-		CollapseUnselectedPockets(pocketItems);
+		Refresh(_source, initialCoinSelection);
+		CollapseUnselectedPockets();
 
-		TreeDataGridSource = new HierarchicalTreeDataGridSource<CoinControlItemViewModelBase>(pocketItems)
+		walletViewModel.UiTriggers.TransactionsUpdateTrigger
+			.Do(_ => Refresh())
+			.Do(_ => RenewSelectionUpdater())
+			.Subscribe()
+			.DisposeWith(_disposables);
+
+		TreeDataGridSource = new HierarchicalTreeDataGridSource<CoinControlItemViewModelBase>(_source)
 		{
 			Columns =
 			{
@@ -40,57 +51,108 @@ public partial class CoinSelectorViewModel : ViewModelBase, IDisposable
 				IndicatorsColumn(),
 				AmountColumn(),
 				AnonymityScoreColumn(),
-				PocketColumn(),
+				PocketColumn()
 			}
 		};
 
-		var coins = GetAllCoinItems(pocketItems);
-
-		coins
-			.AsObservableChangeSet(x => x.SmartCoin)
-			.AutoRefresh(x => x.IsSelected, TimeSpan.FromMilliseconds(100), scheduler: RxApp.MainThreadScheduler)
-			.ToCollection()
-			.Select(x => x.Where(m => m.IsSelected == true))
-			.Select(models => models.Select(x => x.SmartCoin).ToImmutableList())
-			.BindTo(this, model => model.SelectedCoins);
+		TreeDataGridSource.DisposeWith(_disposables);
 	}
 
 	public HierarchicalTreeDataGridSource<CoinControlItemViewModelBase> TreeDataGridSource { get; }
 
 	public void Dispose()
 	{
-		foreach (var pocket in TreeDataGridSource.Items.OfType<PocketCoinControlItemViewModel>())
+		_disposables.Dispose();
+		_selectedCoinsUpdater.Dispose();
+		Dispose(_source);
+	}
+
+	private static void Populate(IList<CoinControlItemViewModelBase> pocketItems, Wallet wallet)
+	{
+		var newItems = wallet
+			.GetPockets()
+			.Select(pocket => new PocketCoinControlItemViewModel(pocket));
+
+		Dispose(pocketItems);
+
+		pocketItems.Clear();
+		pocketItems.AddRange(newItems);
+	}
+
+	private static void Dispose(IEnumerable<CoinControlItemViewModelBase> items)
+	{
+		foreach (var pocket in items.OfType<PocketCoinControlItemViewModel>())
 		{
 			pocket.Dispose();
 		}
-
-		TreeDataGridSource.Dispose();
 	}
 
-	private static IList<CoinCoinControlItemViewModel> GetAllCoinItems(IEnumerable<PocketCoinControlItemViewModel> pockets)
+	private void RenewSelectionUpdater()
 	{
-		return pockets
-			.SelectMany(x => x.Children)
-			.OfType<CoinCoinControlItemViewModel>()
-			.ToList();
+		_selectedCoinsUpdater.Dispose(); // Dispose any previous updater
+
+		_selectedCoinsUpdater = GetAllCoins()
+			.AsObservableChangeSet()
+			.AutoRefresh(x => x.IsSelected)
+			.ToCollection()
+			.Select(list => new ReadOnlyCollection<SmartCoin>(list.Where(item => item.IsSelected == true).Select(x => x.SmartCoin).ToList()))
+			.BindTo(this, x => x.SelectedCoins);
 	}
 
-	private static void SyncSelectedItems(IEnumerable<PocketCoinControlItemViewModel> items, IEnumerable<SmartCoin> selectedCoins)
+	private void CollapseUnselectedPockets()
 	{
-		var allCoins = GetAllCoinItems(items);
-		var selected = allCoins.Where(x => selectedCoins.Any(other => other == x.SmartCoin));
-		foreach (var viewModel in selected)
-		{
-			viewModel.IsSelected = true;
-		}
-	}
-
-	private static void CollapseUnselectedPockets(IReadOnlyCollection<PocketCoinControlItemViewModel> pocketItems)
-	{
-		foreach (var pocket in pocketItems.Where(x => x.IsSelected == false))
+		foreach (var pocket in _source.Where(x => x.IsSelected == false))
 		{
 			pocket.IsExpanded = false;
 		}
+	}
+
+	private IList<SmartCoin> GetSelectedCoins()
+	{
+		return _source
+			.SelectMany(model => model.Children)
+			.Cast<CoinCoinControlItemViewModel>()
+			.Where(x => x.IsSelected == true)
+			.Select(x => x.SmartCoin)
+			.ToList();
+	}
+
+	private List<CoinCoinControlItemViewModel> GetAllCoins()
+	{
+		return _source
+			.SelectMany(model => model.Children)
+			.Cast<CoinCoinControlItemViewModel>()
+			.ToList();
+	}
+
+	private void SyncSelectedItems(IList<SmartCoin> selectedCoins)
+	{
+		var coinItems = _source.SelectMany(x => x.Children).Cast<CoinCoinControlItemViewModel>();
+		foreach (var coinItem in coinItems)
+		{
+			coinItem.IsSelected = selectedCoins.Any(x => x == coinItem.SmartCoin);
+		}
+	}
+
+	private void Refresh()
+	{
+		Refresh(_source, GetSelectedCoins());
+	}
+
+	private void Refresh(IList<CoinControlItemViewModelBase> collection, IList<SmartCoin> selectedItems)
+	{
+		Populate(collection, _wallet);
+		SyncSelectedItems(selectedItems);
+	}
+
+	private static Comparison<TSource?> SortAscending<TSource, TProperty>(Func<TSource, TProperty> selector)
+	{
+		return (x, y) => Comparer<TProperty>.Default.Compare(selector(x!), selector(y!));
+	}
+
+	private static Comparison<TSource?> SortDescending<TSource, TProperty>(Func<TSource, TProperty> selector)
+	{
+		return (x, y) => Comparer<TProperty>.Default.Compare(selector(y!), selector(x!));
 	}
 
 	private static int GetLabelPriority(CoinControlItemViewModelBase coin)
@@ -127,24 +189,7 @@ public partial class CoinSelectorViewModel : ViewModelBase, IDisposable
 
 		return 0;
 	}
-
-	private static IReadOnlyCollection<PocketCoinControlItemViewModel> CreatePocketItems(IEnumerable<Pocket> pockets)
-	{
-		return pockets
-			.Select(pocket => new PocketCoinControlItemViewModel(pocket))
-			.ToList();
-	}
-
-	private static Comparison<TSource?> SortAscending<TSource, TProperty>(Func<TSource, TProperty> selector)
-	{
-		return (x, y) => Comparer<TProperty>.Default.Compare(selector(x!), selector(y!));
-	}
-
-	private static Comparison<TSource?> SortDescending<TSource, TProperty>(Func<TSource, TProperty> selector)
-	{
-		return (x, y) => Comparer<TProperty>.Default.Compare(selector(y!), selector(x!));
-	}
-
+	
 	private static IColumn<CoinControlItemViewModelBase> ChildrenColumn()
 	{
 		return new HierarchicalExpanderColumn<CoinControlItemViewModelBase>(
